pyproject.toml
[tool.poetry]
name = "sat-cluster-api"
version = "0.1.1"
description = ""
authors = ["Sean Crowley <sean.crowley@datirium.com>"]
readme = "README.md"

[tool.poetry.scripts]
start-cluster-api = "sat_cluster_api.__main__:run_app"

[tool.poetry.dependencies]
python = "^3.8.12"
connexion = {extras = ["swagger-ui"], version = "^2.14.2"}
ruamel-yaml = "^0.17.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
